Interaktion angepasst
- Mausklicks sind bis auf den Mine Mode nun nur noch möglich, wenn die Taste zwischendurch losgelassen wird
- Übergänge in die anderen Modes nun aus jedem Mode möglich

+++ VS Projekt/Underlord/Underlord/Underlord/Environment/Map.cs
@@ -67,7 +67,7 @@ namespace Underlord.Environment
         #region Constructor
         public Map(int sidelength, Entity.Vars_Func.HexTyp typ, Boolean newGame, float hexagonSideLength)
         {
-            map = new Hexagon[(int)(sidelength * sidelength)];
+            map = new Hexagon[sidelength * sidelength];
             rooms = new List<Room>();
             nests = new List<Nest>();
             entrances = new List<Nest>();
